Fix MD5 hashing for x86_64

+++ neo/idlib/hashing/MD5.cpp
@@ -257,9 +257,9 @@ void MD5_Final( MD5_CTX *ctx, unsigned char digest[16] ) {
 MD5_BlockChecksum
 ===============
 */
-unsigned long MD5_BlockChecksum( const void *data, int length ) {
-	unsigned long	digest[4];
-	unsigned long	val;
+unsigned int MD5_BlockChecksum( const void *data, int length ) {
+	unsigned int	digest[4];
+	unsigned int	val;
 	MD5_CTX			ctx;
 
 	MD5_Init( &ctx );
