+++ neo/idlib/hashing/MD5.h
@@ -10,6 +10,6 @@
 ===============================================================================
 */
 
-unsigned long MD5_BlockChecksum( const void *data, int length );
+unsigned int MD5_BlockChecksum( const void *data, int length );
 
 #endif /* !__MD5_H__ */
